refractor: moved middleware and business logic for endpoints to corresponding folders and files
