+++ server/server-web.js
@@ -1,10 +1,8 @@
-const { getVehiclesInRange, updateVehicleStatus, getVehicle, getVehicles } = require('./store/vehicles');
-const { getBidsForRequest, deleteBidsForRequest } = require('./store/bids');
-const { getOrCreateUser } = require('./store/users');
-const { createRequest, getRequest, deleteRequest } = require('./store/requests');
-const { createMission, getLatestMissionId, getMission, updateMission } = require('./store/missions');
-const { createMissionUpdate } = require('./store/mission_updates');
-const { hasStore } = require('./lib/environment');
+const cors = require('./middleware/cors');
+const getOrCreateUser = require('./middleware/getOrCreateUser');
+
+const StatusController = require('./controllers/StatusController');
+const RequestController = require('./controllers/RequestController');
 
 // Create thrift connection to Captain
 require('./client-thrift').start({port: process.env.CAPTAIN_PORT, host: process.env.CAPTAIN_HOST});
@@ -13,107 +11,14 @@ const express = require('express');
 const app = express();
 const port = process.env.WEB_SERVER_PORT || 8888;
 
-// Allow CORS
-app.use((req, res, next) => {
-  res.header('Access-Control-Allow-Origin', '*');
-  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
-  next();
-});
+app.use(cors);
+app.use(getOrCreateUser());
 
-// Get or create user
-app.use(async (req, res, next) => {
-  const { user_id } = req.query;
-  let user = await getOrCreateUser(user_id);
-  req.user = user;
-  next();
-});
+app.get('/status', StatusController.getStatus);
 
-// Define routes
-app.get('/', (req, res) => {
-  res.send('hello world');
-});
-
-app.get('/status', async (req, res) => {
-  const { lat, long, requestId, user_id } = req.query;
-  const status = 'idle';
-  const latestMissionId = await getLatestMissionId(user_id);
-  const latestMission = await getMission(latestMissionId);
-  const bids = (!hasStore() || !requestId) ? [] : await getBidsForRequest(requestId);
-  let vehicles = [];
-  if (hasStore()) {
-    if (bids.length > 0) {
-      vehicles = await getVehicles(bids.map(bid => bid.vehicle_id));
-    } else {
-      vehicles = await getVehiclesInRange(
-        { lat: parseFloat(lat), long: parseFloat(long) },
-        7000
-      );
-    }
-  }
-
-  if (latestMission) {
-    switch (latestMission.status) {
-    case 'awaiting_signatures': {
-      let elapsedTime = Date.now() - latestMission.user_signed_at;
-      let elapsedSeconds = ((elapsedTime % 60000) / 1000).toFixed(0);
-      if (elapsedSeconds > 6) {
-        await updateMission(latestMissionId, {'vehicle_signed_at': Date.now(), 'status':'in_progress' });
-        await updateVehicleStatus(latestMission.vehicle_id, 'travelling_pickup');
-        await createMissionUpdate(latestMissionId, 'travelling_pickup');
-      }
-      res.json({status, vehicles, bids});
-      break;
-    }
-    case 'in_progress': {
-      const mission = latestMission;
-      const vehicle = await getVehicle(latestMission.vehicle_id);
-      vehicles = [vehicle];
-      const status = 'in_mission';
-      res.json({status, vehicles, bids, mission});
-      break;
-    }
-    }
-  } else {
-    res.json({ status, vehicles, bids });
-  }
-});
-
-app.get('/request/new', async (req, res) => {
-  const { user_id, pickup, dropoff, requested_pickup_time, size, weight } = req.query;
-  const requestId = await createRequest({
-    user_id, pickup, dropoff, requested_pickup_time, size, weight
-  });
-  if (requestId) {
-    res.json({ requestId });
-  } else {
-    res.status(500).send('Something broke!');
-  }
-});
-
-app.get('/request/cancel', async (req, res) => {
-  const { requestId } = req.query;
-  const request = await getRequest(requestId);
-  if (request) {
-    await deleteRequest(requestId);
-    await deleteBidsForRequest(requestId);
-    res.send('request cancelled');
-  } else {
-    res.status(500).send('Something broke!');
-  }
-});
-
-app.get('/choose_bid', async (req, res) => {
-  const { user_id, bid_id } = req.query;
-  const mission = await createMission({
-    user_id, bid_id
-  });
-  if (mission) {
-    await updateVehicleStatus(mission.vehicle_id, 'contract_received');
-    res.json({ mission });
-  } else {
-    res.status(500).send('Something broke!');
-  }
-});
+app.get('/request/new', RequestController.newRequest);
+app.get('/request/cancel', RequestController.cancelRequest);
+app.get('/choose_bid', RequestController.chooseBid);
 
 module.exports = {
   start: () => {
